Another quick hack

+++ deletetweets/removefollows.py
@@ -12,18 +12,19 @@ class TweetUnfollower(object):
         self.twitter_api = twitter_api
         self.dry_run = dry_run
 
-    def destroy(self, tweet_id):
+    def destroy(self, all, tweet_id):
         try:
             user = self.twitter_api.GetUser(user_id=tweet_id)
 
             print("delete friend %s" % user.screen_name)
             if not self.dry_run:
-                self.twitter_api.DestroyFriendship(user_id=tweet_id)
+                if not all:
+                    self.twitter_api.DestroyFriendship(user_id=tweet_id)
                 self.twitter_api.DestroyMute(user_id=tweet_id)
         except twitter.TwitterError as err:
             print("Exception: %s\n" % err.message)
 
-def delete(dry_run=False):
+def delete(all, dry_run=False):
     api = twitter.Api(consumer_key=os.environ["TWITTER_CONSUMER_KEY"],
                       consumer_secret=os.environ["TWITTER_CONSUMER_SECRET"],
                       access_token_key=os.environ["TWITTER_ACCESS_TOKEN"],
@@ -34,7 +35,11 @@ def delete(dry_run=False):
     mutes = set(api.GetMutesIDs())
     friends = set(api.GetFriendIDs())
 
-    for id in mutes.intersection(friends):
-        destroyer.destroy(id)
+    target = mutes
+    if not all:
+        target = mutes.intersection(friends)
+
+    for id in target:
+        destroyer.destroy(all, id)
 
     sys.exit()
